refactor: Extract RatingInfoBox from Cars component
