+++ src/components/cars/Cars.js
@@ -2,6 +2,7 @@ import React, { useState } from "react";
 import axios from "axios";
 import { Tabs, Tab } from "@mui/material";
 import CarForm from "./CarForm";
+import CarRatings from "./CarRatings";
 
 export default function CarInfo() {
   const [data, setData] = useState({ recalls: [], ratings: [] });
@@ -41,32 +42,6 @@ export default function CarInfo() {
     </div>
   );
 
-  const RatingInfoBox = ({ ratingItem }) => (
-    <div className="border border-gray-300 rounded p-4 my-2">
-      <h2 className="text-lg font-semibold">
-        Vehicle: {ratingItem.VehicleDescription}
-      </h2>
-      <p className="text-base">Overall Rating: {ratingItem.OverallRating}</p>
-      <p className="text-sm">
-        Front Crash Rating: {ratingItem.OverallFrontCrashRating}
-      </p>
-      <p className="text-sm">
-        Side Crash Rating: {ratingItem.OverallSideCrashRating}
-      </p>
-      <p className="text-sm">Rollover Rating: {ratingItem.RolloverRating}</p>
-      <p className="text-sm">
-        Electronic Stability Control:{" "}
-        {ratingItem.NHTSAElectronicStabilityControl}
-      </p>
-      <p className="text-sm">
-        Forward Collision Warning: {ratingItem.NHTSAForwardCollisionWarning}
-      </p>
-      <p className="text-sm">
-        Lane Departure Warning: {ratingItem.NHTSALaneDepartureWarning}
-      </p>
-    </div>
-  );
-
   return (
     <div className="container mx-auto p-5">
       <div
@@ -85,12 +60,7 @@ export default function CarInfo() {
 
         {hasFetchedData && (
           <div className="flex justify-between mt-2">
-            <div className="w-1/2 pr-2">
-              <h2 className="text-lg font-semibold">Rating Information</h2>
-              {data.ratings.map((ratingItem, index) => (
-                <RatingInfoBox key={index} ratingItem={ratingItem} />
-              ))}
-            </div>
+            <CarRatings ratings={data.ratings} />
             <div className="w-1/2 pl-2">
               {data.recalls.length > 0 ? (
                 <>
